src/test/java/garhettM/songr/AlbumTest.java
package garhettM.songr;

import org.junit.Test;
import static org.junit.Assert.*;



public class AlbumTest {


    @Test public void titleTest() {
        Album testMe = new Album("FakeTitle", "FakeArtist", 15, 124, "http:www.fakeimg.com/jpg");

        assertEquals(5, testMe.setSongCount(5));
        assertEquals(5, testMe.getSongCount());
        assertEquals(234, testMe.setLength(234));
        assertEquals(234, testMe.getLength());
        assertEquals("NewFake Artist", testMe.setArtist("NewFake Artist"));
        assertEquals("NewFake Artist", testMe.getArtist());
        assertEquals("NewFake Title", testMe.setArtist("NewFake Title"));
        assertEquals("NewFake Title", testMe.getArtist());
        assertEquals("fakeadress.com/jpg", testMe.setImgUrl("fakeadress.com/jpg"));
        assertEquals("fakeadress.com/jpg", testMe.getImgUrl());
    }

}
